Added a hiccup workflow
Files with ".clj" extensions in the fragments directory are pulled in as
Hiccup files and parsed into HTML. Helper functions to generate more
complicated HTML fragments (like side- & margin-notes, so far) are
applied by eval'ing the supplied Hiccup in the ssgen namespace - e.g.
this is for TRUSTED HICCUP ONLY, not general Hiccup data submitted via
the web (just as an example).