Meta-posts, for real this time
Using a joint path for the URL and the file location on disk proved to
be an awkward if not outright bad idea, so now each article needs to
have a url slug chosen by the user in advance. I probably need to
enforce some sort of policy here (a la the MusicTown setup).

Anyway the upshot is the hiccup files are now served using the meta-post
format. Meta-posts now require the file path and URL for each post to be
set.